feat - finished unit tests for repository layer

+++ Backend/src/main/java/com/firomsa/ecommerce/model/Address.java
@@ -1,13 +1,19 @@
 package com.firomsa.ecommerce.model;
 
-import jakarta.persistence.*;
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "addresses")
@@ -19,7 +25,7 @@ public class Address {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "user_id", referencedColumnName = "id")
@@ -43,7 +49,7 @@ public class Address {
     private String phone;
     @NotNull
     @Builder.Default
-    private Boolean active =  Boolean.FALSE;
+    private Boolean active = Boolean.FALSE;
     @NotNull
     private LocalDateTime createdAt;
     @NotNull
